requirements.txt
# Core Data Science
pandas>=2.1.0
numpy>=1.26.0

# Chemoinformatics
rdkit>=2023.9.1
mordred>=1.2.0

# Machine Learning - Classical
scikit-learn>=1.3.0
xgboost>=1.7.6
scipy>=1.11.0

# Machine Learning - Deep Learning (for GNN)
torch>=2.0.0
pytorch-lightning>=2.0.0
torch-geometric>=2.3.0
ogb>=1.3.6  # Open Graph Benchmark
optuna>=3.0.0  # Bayesian optimization
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
umap-learn>=0.5.0  # For UMAP dimensionality reduction

# Development
jupyter>=1.0.0

# DeepChem - Currently incompatible with Python 3.12
# deepchem>=2.6.0  # Commented out due to Python 3.12 compatibility issues
# Note: utils/featurizers.py requires deepchem - use Classification/src/sider_featurizer.py instead
